api/requirements.txt
amqp==5.0.7
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0
django-cors-headers==3.10.1
django-debug-toolbar==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.13.0
djangorestframework-simplejwt==5.0.0
kombu==5.2.2
mysqlclient==2.1.0
Pillow==8.4.0
prompt-toolkit==3.0.24
PyJWT==2.3.0
pytz==2021.3
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
